+++ frontend/src/main.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import { AuthProvider } from './auth/AuthContext';
+import { SpinLockProvider } from './spin/SpinLockContext';
 import App from './App';
 import { initAmplifyIfNeeded } from './aws-amplify-init';
 import './index.css';
@@ -11,7 +12,9 @@ initAmplifyIfNeeded().finally(() => {
     <React.StrictMode>
       <BrowserRouter>
         <AuthProvider>
-          <App />
+          <SpinLockProvider>
+            <App />
+          </SpinLockProvider>
         </AuthProvider>
       </BrowserRouter>
     </React.StrictMode>